e_library/requirements.txt
django-3.1.7
Pillow
django-crispy-forms
six
taggit
django-filter